fixes: #924 - Windows build fails in volrender.cc
git-svn-id: https://bioproj.extra.cea.fr/neurosvn/brainvisa/anatomist/anatomist-free/branches/3.2@36346 613412ce-1361-4b7a-93bf-a3c75af1bae4

+++ src/volrender/object/volrender.cc
@@ -1124,8 +1124,8 @@ void VolRender::createDefaultPalette( const string & name )
             xrgb.red() = rgb.red();
             xrgb.green() = rgb.green();
             xrgb.blue() = rgb.blue();
-            xrgb.alpha() = (unsigned char) sqrt( sqr( rgb.red() )
-              + sqr( rgb.green() ) + sqr( rgb.blue() ) );
+            xrgb.alpha() = (unsigned char) sqrt( (double)( sqr( rgb.red() )
+              + sqr( rgb.green() ) + sqr( rgb.blue() ) ) );
           }
           theAnatomist->palettes().push_back( apal );
         }
